+++ src/lib/Scenes/City/Components/EventSection/index.tsx
@@ -35,12 +35,11 @@ export class EventSection extends React.Component<Props> {
           <Serif size="8">{title}</Serif>
         </Box>
         {this.renderEvents()}
-        <Box px={2} mb={2}>
-          <CaretButton
-            onPress={() => this.viewAllPressed()}
-            text={data.length === 1 ? "View 1 show" : `View all ${data.length} shows`}
-          />
-        </Box>
+        {data.length > 2 && (
+          <Box px={2} mb={2}>
+            <CaretButton onPress={() => this.viewAllPressed()} text={`View all ${data.length} shows`} />
+          </Box>
+        )}
       </>
     )
   }